requirements.txt
PyQt5
numpy
matplotlib
scikit-fuzzy
coppeliasim